+++ contracts/zero-ex/contracts/src/transformers/bridges/mixins/MixinVelodromeV2.sol
@@ -1,0 +1,69 @@
+// SPDX-License-Identifier: Apache-2.0
+/*
+  Copyright 2023 ZeroEx Intl.
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+*/
+
+pragma solidity ^0.6.5;
+pragma experimental ABIEncoderV2;
+
+import "@0x/contracts-erc20/src/v06/LibERC20TokenV06.sol";
+import "@0x/contracts-erc20/src/IERC20Token.sol";
+
+interface IVelodromeV2Router {
+    struct Route {
+        address from;
+        address to;
+        bool stable;
+        address factory;
+    }
+
+    /// @notice Swap one token for another
+    /// @param amountIn     Amount of token in
+    /// @param amountOutMin Minimum amount of desired token received
+    /// @param routes       Array of trade routes used in the swap
+    /// @param to           Recipient of the tokens received
+    /// @param deadline     Deadline to receive tokens
+    /// @return amounts     Array of amounts returned per route
+    function swapExactTokensForTokens(
+        uint256 amountIn,
+        uint256 amountOutMin,
+        Route[] calldata routes,
+        address to,
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
+}
+
+contract MixinVelodromeV2 {
+    using LibERC20TokenV06 for IERC20Token;
+
+    function _tradeVelodromeV2(
+        IERC20Token sellToken,
+        uint256 sellAmount,
+        bytes memory bridgeData
+    ) internal returns (uint256 boughtAmount) {
+        (IVelodromeV2Router router, IVelodromeV2Router.Route[] memory routes) = abi.decode(
+            bridgeData,
+            (IVelodromeV2Router, IVelodromeV2Router.Route[])
+        );
+        sellToken.approveIfBelow(address(router), sellAmount);
+
+        uint256[] memory amounts = router.swapExactTokensForTokens(
+            sellAmount,
+            1,
+            routes,
+            address(this),
+            block.timestamp + 1
+        );
+
+        return amounts[amounts.length - 1];
+    }
+}